restructured components into folders,. installed style-loader and css-loader dependencies, added css file for each component folder, renamed project-item and project-items components
