+++ src/components/layout/feature-panel.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './layout.css';
 
 export default class FeaturePanel extends React.Component {
     constructor(props){
