+++ src/components/layout/footer.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './layout.css';
 
 export default class Footer extends React.Component {
     constructor(props){
