+++ src/components/layout/header.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './layout.css';
 
 export default class Header extends React.Component {
     constructor(props){
