+++ src/components/layout/hero-panel.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './layout.css';
 
 export default class HeroPanel extends React.Component {
     constructor(props){
