+++ src/components/layout/layout.css
@@ -1,0 +1,25 @@
+/* header and nav styling */
+.header-bg {
+    background-color: #4A306D;
+    color: white;
+}
+.menu-nav > * {
+    margin-left: 10px;
+    cursor: pointer;
+}
+/* hero-panel component styling */
+.hero-bg {
+    background: url('https://bit.ly/2Kj2doT') no-repeat;
+    background-size: cover;
+    height: 600px;
+}
+.hero-content {
+    background-color: rgba(255, 255, 255, 0.6);
+}
+/* feature-panel component styling */
+.feature-bg {
+    padding: 60px 0;
+}
+.feature-icon{
+    max-width: 125px;
+}
