+++ src/components/project/project-card.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './project.css';
 
 export default class ProjectCard extends React.Component{
     handleProjectClick(event) {
