+++ src/components/project/timeline-entry.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './project.css';
 
 export default class TimelineEntry extends React.Component {
     render() {
